requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
huggingface_hub
sentence-transformers
PyPDF2
python-dotenv
